refactor: pageのcss置き換え
