+++ .vscode/settings.json
@@ -2,6 +2,7 @@
   "cSpell.words": [
     "clsx",
     "cornflowerblue",
+    "Laravel",
     "lightgray",
     "NAVI",
     "tocs",
