+++ src/app/article/[id]/Article.css.ts
@@ -1,0 +1,7 @@
+import { style } from '@vanilla-extract/css';
+
+export const styles = {
+  articleContent: style({
+    margin: 10
+  })
+};
