--- src/app/article/[id]/Article.module.css
@@ -1,3 +1,0 @@
-.articleContent {
-  margin: 10px;
-}
