--- src/app/article/[id]/Article.module.css.d.ts
@@ -1,4 +1,0 @@
-declare const styles: {
-  readonly articleContent: string;
-};
-export = styles;
