--- src/app/categories/Categories.module.css
@@ -1,4 +1,0 @@
-.baseHeading1Categories {
-  padding: 0 0 10px;
-  border-bottom: 2.5px solid lightgray;
-}
