--- src/app/tags/Tags.module.css
@@ -1,4 +1,0 @@
-.baseHeading1Tags {
-  padding: 0 0 10px;
-  border-bottom: 2.5px solid lightgray;
-}
